wallet: move RenameAccount across all backends
