+++ wallet/internal/db/pg/accountstore_renameaccount.go
@@ -1,0 +1,61 @@
+package pg
+
+import (
+	"context"
+
+	"github.com/btcsuite/btcwallet/wallet/internal/db"
+	"github.com/btcsuite/btcwallet/wallet/internal/sql/pg/sqlc"
+)
+
+// RenameAccount changes the name of an account. The account can be identified
+// by its old name or its account number.
+func (s *Store) RenameAccount(ctx context.Context,
+	params db.RenameAccountParams) error {
+
+	return s.execWrite(ctx, func(qtx *sqlc.Queries) error {
+		renameQueries := accountRenameQueries{q: qtx}
+
+		return db.RenameAccountByQuery(
+			ctx, params, renameQueries.byNumber, renameQueries.byName,
+		)
+	})
+}
+
+// accountRenameQueries groups PostgreSQL account rename query methods.
+type accountRenameQueries struct {
+	q *sqlc.Queries
+}
+
+// byNumber renames an account identified by wallet ID, scope, and account
+// number.
+func (p accountRenameQueries) byNumber(ctx context.Context,
+	params db.RenameAccountParams) error {
+
+	return db.RenameAccount(
+		ctx, p.q.UpdateAccountNameByWalletScopeAndNumber,
+		sqlc.UpdateAccountNameByWalletScopeAndNumberParams{
+			NewName:       params.NewName,
+			WalletID:      int64(params.WalletID),
+			Purpose:       int64(params.Scope.Purpose),
+			CoinType:      int64(params.Scope.Coin),
+			AccountNumber: db.NullableUint32ToSQLInt64(params.AccountNumber),
+		}, params,
+	)
+}
+
+// byName renames an account identified by wallet ID, scope, and old account
+// name.
+func (p accountRenameQueries) byName(ctx context.Context,
+	params db.RenameAccountParams) error {
+
+	return db.RenameAccount(
+		ctx, p.q.UpdateAccountNameByWalletScopeAndName,
+		sqlc.UpdateAccountNameByWalletScopeAndNameParams{
+			NewName:  params.NewName,
+			WalletID: int64(params.WalletID),
+			Purpose:  int64(params.Scope.Purpose),
+			CoinType: int64(params.Scope.Coin),
+			OldName:  params.OldName,
+		}, params,
+	)
+}
